+++ tests/testthat/test_execute.R
@@ -63,4 +63,13 @@ describe("execute", {
         output <- scriptexec::execute("dir", runner = runner)
         expect_equal(output$status, 0)
     })
+    
+    it("error during invocation", {
+        expect_error(scriptexec::execute("badcommand"))
+    })
+    
+    it("print_commands", {
+        output <- scriptexec::execute("dir", print_commands = TRUE)
+        expect_equal(output$status, 0)
+    })
 })